[BE] 소셜 로그인 깃헙 추가

Closes #150

See merge request s07-final/S07P31B205!192

+++ frontend/src/components/welcome/OAuth2RedirectHandler.tsx
@@ -2,9 +2,17 @@ import { useEffect } from "react";
 
 export const OAuth2RedirectHandler = () => {
   useEffect(() => {
-    const token = window.location.search.substring(7) as string;
-    localStorage.clear();
-    localStorage.setItem("token", token);
+    function searchParam(key: string) {
+      return new URLSearchParams(window.location.search).get(key);
+    }
+    const token = searchParam("token");
+    const error = searchParam("error");
+    if (token) {
+      localStorage.clear();
+      localStorage.setItem("token", token);
+    } else if (error) {
+      alert(error);
+    }
     window.location.replace("/");
   }, []);
   return <div></div>;
